fstadv

Extrapolate data from FORECAST/ADVISORY products.

fstadv(link, msg = FALSE)

Arguments

link
URL of a specific FORECAST/ADVISORY product

msg
Display each link as being worked; default is FALSE

Value

Dataframe

Details

Given a direct link to a forecast/advisory product, parse and return dataframe of values.

Status

Classification of storm, e.g., Tropical Storm, Hurricane, etc.

Name

Name of storm

Adv

Advisory Number

Date

Date of advisory issuance

Key

Unique identifier for the storm

Lat

Latitude, location of the storm. Positive is northern hemisphere. Negative is southern hemisphere.

Lon

Longitude, location of the storm. Positive is eastern hemisphere. Negative is western hemisphere.

Wind

Current wind speed in knots.

Gust

Current maximum wind gusts in knots.

Pressure

Central barometric pressure in millibars.

PosAcc

Position accuracy of storm in nautical miles.

FwdDir

Forward direction of storm on a compass scale (0-359 or NA).

FwdSpeed

Forward speed of storm in knots or NA if stationary/drifting.

Eye

Size of the eye in nautical miles, if available, or NA.

See also

get_fstadv